Add problem status icon to ProblemList.vue
